make getCookie Function
