+++ frontend/src/Components/Util/Cookie.js
@@ -1,0 +1,11 @@
+export const getCookie = (key) => {
+  let value = "NO_HAVE";
+  const cookies = document.cookie.split(";");
+
+  for (let i in cookies) {
+    if (cookies[i].search(key) !== -1) {
+      value = cookies[i].split("=")[1];
+    }
+  }
+  return value;
+};
